Global refactor
